Refactor MongoDB configuration for Product Service
This commit introduces the following changes to the codebase:

- Renamed the `mongodb` service to `mongo` in `docker-compose.yaml`
- Changed the container name from `product-service` to `mongo-product-service` in `docker-compose.yaml`
- Changed the volume mapping from `data:/data` to `./mongo-product-service:/data/db` in `docker-compose.yaml`
- Removed the `mongo-express` service from `docker-compose.yaml`
- Removed the `networks` section from `docker-compose.yaml`
- Added `mongo-product-service` volume to the `.gitignore` file
- Changed the `spring.data.mongodb.username` and `spring.data.mongodb.password` properties to static values (`root` and `root`) in `application-dev.properties`
- Added the `restart: unless-stopped` directive to the `mongo` service in `docker-compose.yaml`

Purpose:
The purpose of this commit is to refactor the MongoDB configuration specifically for the Product Service. The changes include renaming the MongoDB service and container for better clarity, updating the volume mapping to persist data on the host, removing the `mongo-express` service, ensuring the MongoDB container restarts automatically unless explicitly stopped, and excluding mongo-product-service from being tracked by Git. Additionally, the MongoDB username and password properties in `application-dev.properties` have been updated to static values for development purposes.